+++ lib/src/models/ibg_feature_flag.dart
@@ -1,0 +1,6 @@
+class IBGFeatureFlag {
+  String name;
+  String? variant;
+
+  IBGFeatureFlag({required this.name, this.variant});
+}